# Speed up plots, remove obsolete code (SCP-3134)