--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -22,7 +22,6 @@ module.exports = {
         'morpheus': 'readonly',
         'igv': 'readonly',
         'Ideogram': 'readonly',
-        'getViolinProps': 'readonly',
         'ga': 'readonly',
         'bamAndBaiFiles': 'readonly'
     },