--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -20,8 +20,6 @@ module.exports = {
         'd3': 'readonly',
         'Spinner': 'readonly',
         'morpheus': 'readonly',
-        'igv': 'readonly',
-        'Ideogram': 'readonly',
         'ga': 'readonly',
         'bamAndBaiFiles': 'readonly'
     },