--- a/app/assets/javascripts/application.js
+++ b/app/assets/javascripts/application.js
@@ -164,41 +164,6 @@ function elementVisible(element) {
     return $(element).is(":visible");
 }
 
-function paginateStudies(totalPages, order, searchString, project) {
-
-    var target = document.getElementById("pagination");
-    var spinner = new Spinner(paginationOpts).spin(target);
-    var page = parseInt($($(".study-panel").slice(-1)[0]).attr("data-page")) + 1;
-    var dataParams = {};
-    dataParams["page"] = page;
-    if (order !== "") {
-        dataParams["order"] = order;
-    }
-    if (searchString !== "") {
-        dataParams["search_terms"] = searchString;
-    }
-    if (project !== "") {
-        dataParams["scpbr"] = project;
-    }
-    $("#pagination").fadeOut("fast", function() {
-            $.ajax({
-                url: "/single_cell",
-                data: dataParams,
-                dataType: "script",
-                type: "GET",
-                success: function(data){
-                    spinner.stop();
-                    if ( dataParams["page"] < totalPages ) {
-                        $("#pagination").fadeIn("fast");
-                        $(window).bind('scroll', bindScroll);
-                    }
-                }
-            });
-        }
-    );
-}
-
-
 // used for keeping track of position in wizard
 var completed = {
     initialize_expression_form_nav: false,
@@ -366,30 +331,6 @@ var stickyOptions = {
     topPadding: 85
 };
 
-// eweitz 2018-02-08: We removed the 'view-fullscreen' button that calls this last month.  Do we still need this?
-// toggle the Search/View options panel
-function toggleSearch() {
-    $('#search-target').toggleClass('col-md-3 hidden');
-    $('#render-target').toggleClass('col-md-9 col-md-12');
-    $('#search-options-panel').toggleClass('hidden');
-    $('#show-search-options').toggleClass('hidden');
-    if ( $('#show-search-options').css('display') === 'none' ) {
-        $('#show-search-options').tooltip('hide');
-    }
-
-    // trigger resizeEnd to re-render Plotly to use available space
-    $(window).trigger('resize');
-    if ($('#panel-selection').is(':visible')) {
-        if ($('#search-target').is(":visible")) {
-            $('#search-parent').stickyPanel(stickyOptions)
-        } else {
-            if ($('#search-parent').data("stickyPanel.state") !== 'undefined') {
-                $('#search-parent').stickyPanel('unstick')
-            }
-        }
-    }
-}
-
 // functions to show loading modals with spinners
 // callback function will execute after modal completes opening
 function launchModalSpinner(spinnerTarget, modalTarget, callback) {
@@ -457,72 +398,6 @@ function showMessageModal(notice=null, alert=null) {
     }
 }
 
-// Propagate changes from the View Options sidebar to the Search Genes form.
-function updateSearchGeneParams() {
-
-  // Get values from control elements in View Options sidebar
-  var cluster = $('#cluster').val();
-  var annotation = $('#annotation').val();
-  var consensus = $('#search_consensus').val();
-  var subsample = $('#subsample').val();
-  var plot_type = $('#plot_type').val() === undefined ? 'violin' : $('#plot_type').val();
-  var jitter = $('#jitter').val() === undefined ? 'all' : $('#jitter').val();
-  var boxpoints = $('#boxpoints_select').val() === undefined ? 'all' : $('#boxpoints_select').val();
-  var heatmap_size = $('#heatmap');
-  var heatmap_row_centering = $('#heatmap_row_centering').val();
-  var heatmap_size = parseInt($('#heatmap_size').val());
-
-  // These 'search_foo' values exist in hidden form elements in '#search-genes-input'
-  $("#search_cluster").val(cluster);
-  $("#search_annotation").val(''); // clear value first
-  $("#search_annotation").val(annotation);
-  $('#search_plot_type').val(plot_type);
-  $('#search_jitter').val(jitter);
-  $('#search_boxpoints').val(boxpoints);
-  $('#search_heatmap_row_centering').val(heatmap_row_centering);
-  $('#search_heatmap_size').val(heatmap_size);
-  $('#search_subsample').val(subsample);
-}
-
-
-// Gets URL parameters needed for each "render" call, e.g. no-gene, single-gene, multi-gene
-function getRenderUrlParams() {
-
-  // Get values from control elements in View Options sidebar
-  var cluster = $('#cluster').val();
-  var annotation = $('#annotation').val();
-  var consensus = $('#search_consensus').val();
-  var subsample = $('#subsample').val();
-  var plot_type = $('#plot_type').val() === undefined ? 'violin' : $('#plot_type').val();
-  var boxpoints = $('#boxpoints_select').val() === undefined ? 'all' : $('#boxpoints_select').val();
-  var heatmap_row_centering = $('#heatmap_row_centering').val();
-  var heatmap_size = parseInt($('#heatmap_size').val());
-  var color_profile = $('#colorscale').val();
-
-  var urlParams =
-    'cluster=' + cluster +
-    '&annotation=' + annotation +
-    '&boxpoints=' + boxpoints +
-    '&consensus=' + consensus +
-    '&subsample=' + subsample +
-    '&plot_type=' + plot_type +
-    '&heatmap_row_centering=' + heatmap_row_centering +
-    '&heatmap_size=' + heatmap_size +
-    '&colorscale=' + color_profile;
-
-  urlParams = urlParams.replace('%', '%25');
-
-  return urlParams;
-}
-
-// Handles changes in View Options for 'Heatmap' view
-$(document).on('change', '#heatmap_row_centering, #annotation', function() {
-  updateSearchGeneParams();
-});
-$(document).on('click', '#resize-heatmap', function() {
-  updateSearchGeneParams();
-});
-
 // default title font settings for axis titles in plotly
 var plotlyTitleFont = {
     family: 'Helvetica Neue',
@@ -538,14 +413,6 @@ var plotlyLabelFont = {
 };
 
 var plotlyDefaultLineColor = 'rgb(40, 40, 40)';
-
-// default scatter plot colors, a combination of colorbrewer sets 1-3 with tweaks to the yellow members
-//
-// Only use this in legacy ERB templates.  For new JS, use
-// `getColorBrewerColor`a standard convenience method exported from lib/plot.js
-var colorBrewerSet = ["#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#a65628", "#f781bf", "#999999",
-    "#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854", "#ffd92f", "#e5c494", "#b3b3b3", "#8dd3c7",
-    "#bebada", "#fb8072", "#80b1d3", "#fdb462", "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd", "#ccebc5", "#ffed6f"];
 
 // clear out text area in a form
 function clearForm(target) {
@@ -602,27 +469,6 @@ function toggleFastqFields(target, state) {
     nameField.parent().effect('highlight', 1200);
     fastqField.effect('highlight', 1200);
 }
-
-// function to toggle all traces in a Plotly div
-function togglePlotlyTraces(div) {
-    console.log('toggling all traces in ' + div);
-    var plotlyData = document.getElementById(div).data;
-    var visibility = plotlyData[0].visible;
-
-    // if visibility is undefined or true, that means it is visible and we want to set this to 'legendonly'
-    // when visibility === 'legendonly', we can set this back to true to show all traces
-    if( visibility === undefined || visibility === true) {
-        visibility = 'legendonly';
-    } else {
-        visibility = true
-    }
-
-    Plotly.restyle(div, 'visible', visibility);
-    // toggle class of toggle glyph
-    $('#toggle-traces').children().toggleClass('fa-toggle-on fa-toggle-off');
-    console.log('toggle complete in ' + div + '; visibility now ' + visibility);
-}
-
 
 // function to return a plotly histogram data object from an array of input values
 function formatPlotlyHistogramData(valuesHash, offset) {
@@ -844,13 +690,6 @@ function gatherFilesByType(fileType) {
     return matchingfiles;
 }
 
-// garbage collector to clear the search animation on global gene search (in case no results are found)
-window.clearGeneSearchLoading = function() {
-    console.log('Clearing global gene search message');
-    $('#wrap').data('spinner').stop();
-    $('#gene-search-results-count').html($('.gene-panel').length);
-};
-
 // force login on ajax 401
 $(document).ajaxError(function (e, xhr, settings) {
     if (xhr.status === 401) {
@@ -866,26 +705,6 @@ function reopenUiTab(navTarget) {
     if (tab !== '') {
         $(navTarget + ' a[href="' + tab + '"]').tab('show');
     }
-}
-
-// re-render a plot after a user selects a new cluster from the dropdown menu, usually called from a complete() callback
-// in an $.ajax() function
-function renderWithNewCluster(updateStatusText, renderCallback, setAnnotation=true) {
-    if (updateStatusText === 'success') {
-        if (setAnnotation) {
-            var an = $('#annotation').val();
-            $('#search_annotation').val(an);
-            $('#gene_set_annotation').val(an);
-        }
-        renderCallback();
-    }
-}
-
-// extract out identifier for global gene search by trimming off last -cluster, -annotation, or -subsample
-function extractIdentifierFromId(domId) {
-    var idParts = domId.split('-');
-    idParts.pop();
-    return idParts.join('-');
 }
 
 // append a list of options to a select menu dynamically when file_type is changed on sync_study view