--- a/app/controllers/api/v1/visualization/clusters_controller.rb
+++ b/app/controllers/api/v1/visualization/clusters_controller.rb
@@ -212,7 +212,7 @@ module Api
         def self.get_selected_subsample_threshold(param, cluster)
           subsample = nil
           if param.blank?
-            # default to largest threshold available that is <10K
+            # default to largest threshold available that is < 10K
             subsample = ClusterVizService.default_subsampling(cluster)
           elsif param == 'all'
             subsample = nil