--- a/app/controllers/api/v1/visualization/explore_controller.rb
+++ b/app/controllers/api/v1/visualization/explore_controller.rb
@@ -86,9 +86,6 @@ module Api
             geneLists: @study.precomputed_scores.pluck(:name),
             annotationList: AnnotationVizService.get_study_annotation_options(@study, current_api_user),
             clusterGroupNames: ClusterVizService.load_cluster_group_options(@study),
-            # spatialGroupNames is for legacy compatibility -- it should be removed once
-            # the react refactor is no longer feature-flagged
-            spatialGroupNames: spatial_group_options.map { |opt| opt[:name] },
             spatialGroups: spatial_group_options,
             clusterPointAlpha: @study.default_cluster_point_alpha,
             colorProfile: @study.default_color_profile