… app/assets/javascripts/igv.min.map
… app/assets/javascripts/scp-dot-plot.js
… app/assets/javascripts/scp-heatmap.js
… app/assets/javascripts/scp-igv.js
… app/assets/javascripts/scp-infercnv-ideogram.js
… app/assets/javascripts/scp-related-genes-ideogram.js
… app/javascript/components/explore/ExploreDisplayTabs.js
… app/javascript/components/explore/ExploreView.js
… app/javascript/components/explore/GenomeView.js
… app/javascript/components/visualization/DotPlot.js
… app/javascript/components/visualization/RelatedGenesIdeogram.js
… app/javascript/lib/plot.js
… app/javascript/lib/scatter-plot.js
… app/javascript/lib/scp-api.js
… app/javascript/lib/study-overview/explore-default.js
… app/javascript/lib/study-overview/explore-single.js
… app/javascript/lib/study-overview/explore.js
… app/javascript/lib/study-overview/user-annotations.js
… app/javascript/lib/study-overview/view-options.js
… app/javascript/lib/user-auth-tokens.js
… app/javascript/lib/violin-plot.js
… app/javascript/packs/application.js
… app/lib/cluster_viz_service.rb
… app/views/layouts/application.html.erb
… app/views/site/_boxpoints_picker.html.erb
… app/views/site/_colorscale_picker.html.erb
… app/views/site/_expression_annotation_plots_plotly.html.erb
… app/views/site/_expression_annotation_plots_view.erb
… app/views/site/_expression_annotation_scatter_layout.html.erb
… app/views/site/_expression_box_layout.html.erb